customize AppBar and Drawer
